move dataset, model, and utils to separate dir for detection example
